+++ src/VisualStudio/Core/Def/EditorConfigSettings/Formatting/ViewModel/TabSizeSettings.cs
@@ -1,0 +1,18 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.VisualStudio.LanguageServices.EditorConfigSettings.Formatting.ViewModel
+{
+    internal enum TabSizeSettings
+    {
+        _1 = 1,
+        _2,
+        _3,
+        _4,
+        _5,
+        _6,
+        _7,
+        _8,
+    }
+}
